+++ c/array-tiling/pnr3.c
@@ -2,7 +2,7 @@ extern void __VERIFIER_error() __attribute__ ((__noreturn__));
 extern void __VERIFIER_assume(int);
 void __VERIFIER_assert(int cond) { if(!(cond)) { ERROR: __VERIFIER_error(); } }
 extern int __VERIFIER_nondet_int(void);
-void *malloc(unsigned long size);
+void *malloc(unsigned int size);
 
 int ReadFromPort()
 {
