+++ c/array-tiling/rewrev.c
@@ -2,7 +2,7 @@ extern void __VERIFIER_error() __attribute__ ((__noreturn__));
 extern void __VERIFIER_assume(int);
 void __VERIFIER_assert(int cond) { if(!(cond)) { ERROR: __VERIFIER_error(); } }
 extern int __VERIFIER_nondet_int(void);
-void *malloc(unsigned long size);
+void *malloc(unsigned int size);
 
 const int MAX = 100000;
 int SIZE;
